+++ frontend/src/components/Pages/InspectionReportPage/InspectionView.tsx
@@ -1,5 +1,5 @@
 import { Icon, Typography } from '@equinor/eds-core-react'
-import { Task, TaskStatus } from 'models/Task'
+import { Task } from 'models/Task'
 import { useInstallationContext } from 'components/Contexts/InstallationContext'
 import { Icons } from 'utils/icons'
 import { useLanguageContext } from 'components/Contexts/LanguageContext'
@@ -12,18 +12,11 @@ import {
     StyledDialogContent,
     StyledDialogHeader,
     StyledDialogInspectionView,
-    StyledImageCard,
-    StyledImagesSection,
     StyledInfoContent,
     StyledInspection,
-    StyledInspectionCards,
-    StyledInspectionContent,
-    StyledInspectionData,
-    StyledInspectionImage,
-    StyledSection,
 } from './InspectionStyles'
-import { BackendAPICaller } from 'api/ApiCaller'
-import { useQuery } from '@tanstack/react-query'
+import { InspectionOverview } from './InspectionOverview'
+import { fetchImageData } from './InspectionReportUtilities'
 
 interface InspectionDialogViewProps {
     selectedTask: Task
@@ -85,93 +78,11 @@ export const InspectionDialogView = ({ selectedTask, tasks }: InspectionDialogVi
                                     )}
                                 </StyledBottomContent>
                             </div>
-                            <InspectionsViewSection tasks={tasks} dialogView={true} />
+                            <InspectionOverview tasks={tasks} dialogView={true} />
                         </StyledDialogInspectionView>
                     </StyledDialogContent>
                 </StyledDialog>
             )}
         </>
     )
-}
-
-interface InspectionsViewSectionProps {
-    tasks: Task[]
-    dialogView?: boolean | undefined
-}
-
-export const InspectionsViewSection = ({ tasks, dialogView }: InspectionsViewSectionProps) => {
-    const { TranslateText } = useLanguageContext()
-    const { switchSelectedInspectionTask } = useInspectionsContext()
-
-    return (
-        <StyledSection
-            style={{
-                width: dialogView ? '350px' : 'auto',
-                borderColor: dialogView ? 'white' : 'auto',
-                padding: dialogView ? '0px' : 'auto',
-                maxHeight: dialogView ? '60vh' : 'auto',
-            }}
-        >
-            {!dialogView && <Typography variant="h4">{TranslateText('Last completed inspection')}</Typography>}
-            <StyledImagesSection>
-                <StyledInspectionCards>
-                    {Object.keys(tasks).length > 0 &&
-                        tasks.map(
-                            (task) =>
-                                task.status === TaskStatus.Successful && (
-                                    <StyledImageCard
-                                        key={task.isarTaskId}
-                                        onClick={() => switchSelectedInspectionTask(task)}
-                                    >
-                                        <GetInspectionImage task={task} />
-                                        <StyledInspectionData>
-                                            {task.tagId && (
-                                                <StyledInspectionContent>
-                                                    <Typography variant="caption">
-                                                        {TranslateText('Tag') + ':'}
-                                                    </Typography>
-                                                    <Typography variant="body_short">{task.tagId}</Typography>
-                                                </StyledInspectionContent>
-                                            )}
-                                            {task.endTime && (
-                                                <StyledInspectionContent>
-                                                    <Typography variant="caption">
-                                                        {TranslateText('Timestamp') + ':'}
-                                                    </Typography>
-                                                    <Typography variant="body_short">
-                                                        {formatDateTime(task.endTime!, 'dd.MM.yy - HH:mm')}
-                                                    </Typography>
-                                                </StyledInspectionContent>
-                                            )}
-                                        </StyledInspectionData>
-                                    </StyledImageCard>
-                                )
-                        )}
-                </StyledInspectionCards>
-            </StyledImagesSection>
-        </StyledSection>
-    )
-}
-
-const fetchImageData = (task: Task) => {
-    const data = useQuery({
-        queryKey: ['fetchInspectionData', task.isarTaskId],
-        queryFn: async () => {
-            const imageBlob = await BackendAPICaller.getInspection(task.inspection.isarInspectionId)
-            return URL.createObjectURL(imageBlob)
-        },
-        retryDelay: 60 * 1000, // Waits 1 min before retrying, regardless of how many retries
-        staleTime: 10 * 60 * 1000, // If data is received, stale time is 10 min before making new API call
-        enabled:
-            task.status === TaskStatus.Successful &&
-            task.isarTaskId !== undefined &&
-            task.inspection.isarInspectionId !== undefined,
-    })
-    return data
-}
-
-
-const GetInspectionImage = ({ task }: { task: Task }) => {
-    const { data } = fetchImageData(task)
-    return <>{data !== undefined && <StyledInspectionImage src={data} />}</>
 }
